% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.parse_cell_label}
\alias{.parse_cell_label}
\title{Parse one cell label into 0-based (a_level, b_level)}
\usage{
.parse_cell_label(label, factor_a, factor_b, sep)
}
\description{
Parse one cell label into 0-based (a_level, b_level)
}
\keyword{internal}
