% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{grin_criterion}
\alias{grin_criterion}
\title{Combine stopping targets into a criterion}
\usage{
grin_criterion(targets, combine = "all")
}
\arguments{
\item{targets}{A list of \code{grin_target} objects (see \code{\link[=grin_target_precision]{grin_target_precision()}},
\code{\link[=grin_target_probability]{grin_target_probability()}}).}

\item{combine}{\code{"all"} (default, stop only once every target is met) or \code{"any"}.}
}
\description{
Combine stopping targets into a criterion
}
