% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-group.R
\name{grin_plot_params_group}
\alias{grin_plot_params_group}
\title{Plot the distribution of each parameter across many participants}
\usage{
grin_plot_params_group(results, ids = NULL)
}
\arguments{
\item{results}{A list of \code{grin_inference} objects (see \code{\link[=grin_tidy]{grin_tidy()}}).}

\item{ids}{Optional participant IDs; see \code{\link[=grin_tidy]{grin_tidy()}}.}
}
\value{
A ggplot object.
}
\description{
Plot the distribution of each parameter across many participants
}
