% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PARAM_GROUPS}
\alias{PARAM_GROUPS}
\title{Named parameter groups (1-indexed positions into PARAM_NAMES)}
\format{
An object of class \code{list} of length 3.
}
\usage{
PARAM_GROUPS
}
\description{
Named parameter groups (1-indexed positions into PARAM_NAMES)
}
\keyword{datasets}
